// src/shared/interface.ts
export interface GatheredUtxos {
  utxos: FormattedUtxo[]
  totalAmount: number
}

export interface FormattedUtxo {
  txId: string
  outputIndex: number
  satoshis: number
  scriptPk: string
  address: string
  inscriptions: any[]
  confirmations: number
}
// ...
export interface AlkanesPayload {
  body: Uint8Array
  cursed: boolean
  tags: { contentType: string }
}

export interface AlkaneId {
  block: string
  tx: string
}